py/pyproject.toml
[project]
name = "blanket-lm"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "beautifulsoup4>=4.13.4",
    "google-genai>=1.14.0",
    "markdown>=3.8",
    "pymupdf4llm>=0.0.24",
    "requests>=2.32.3",
]

[project.scripts]
# package.module:func
blkmkd= "blanketlm:markdown"
blkcly= "blanketlm:classify"
blksmry= "blanketlm:summary"
blklinks = "blanketlm:extract" 
lbkretrieve = "blanketlm:retrieve"
blkfinetune = "blanketlm:finetune"
blkpdfmd = "blanketlm:pdfmd"
[tool.pyright]

[tool.setuptools]
package-dir = {"" = "src"}

[tool.uv]
package = true

[dependency-groups]
dev = [
    "black>=25.1.0",
    "csvkit>=2.1.0",
    "ipython>=9.2.0",
    "pyright>=1.1.400",
]
